Only export circ_lib in 852 subfield b
Aspen can only handle one of these at a time, and my hunch is we want
to send them the asset.copy.circ_lib.

This should be implemented as an option similar to that requested for
Evergreen's Perl marc_export in Lp 2004587: multiple $bs in items from
marc_export can confuse external systems.

This is a quick and dirty hack to get some data out to ByWater in a
reasonable time.

Signed-off-by: Jason Stephenson <jstephenson@cwmars.org>